BREAKING: deprecate `getViewData()` in favour of loadSession
=== why? ===

For the repeater (and other context-sensitive rendering), we will need
to access and mutate view data through `res.locals`. This means that all
our data needs to be available there. So, instead of passing in data to
`res.render(...)`, we consistently use a middleware that loads data into
`res.locals`. This results in easier-to-follow code, and it is clearer
that the choice of what data to load comes from the schema. For locals
that don't appear in the schema, implementers can use
`loadKeys(...keys)` or `loadFullSession` to get data from the session,
or implement a custom middleware that adds data to `res.locals`.

For the vast majority of cases, rendering should be done with the
middleware `route.render()`.