Add explicit max-height to svg container
On IE on Windows 7 and under, the height of the SVG in the header
is far too large, even though it visually is the right size.
(that is, there is a lot of whitespace underneath the img element.)

Checked how Canada.ca solves this problem and they set a max-height
at 40px, which seems to work and isn't too complicated.
Looked up some stackoverflow questions but they mostly had
complicated fixes so gonna go with this instead.