The repeater! (#132)
* enable extended body parsing for nested data

* BREAKING: deprecate `getViewData()` in favour of loadSession

=== why? ===

For the repeater (and other context-sensitive rendering), we will need
to access and mutate view data through `res.locals`. This means that all
our data needs to be available there. So, instead of passing in data to
`res.render(...)`, we consistently use a middleware that loads data into
`res.locals`. This results in easier-to-follow code, and it is clearer
that the choice of what data to load comes from the schema. For locals
that don't appear in the schema, implementers can use
`loadKeys(...keys)` or `loadFullSession` to get data from the session,
or implement a custom middleware that adds data to `res.locals`.

For the vast majority of cases, rendering should be done with the
middleware `route.render()`.

* add context middleware primitives for data

`enterContext` and `exitContext` allow us to dynamically traverse nested
data in the session. Using these, we can change the "context" from which
we get data and errors, and automatically generate `name` attributes
appropriate for the nested data.

* update all the control macros to use getData(...) etc

this way, the data, error, and name will all be relative to the context.

* use loadFullSession in the confirmation page

* remove the data.* scope in the confirmation page

* use the new style controls in the personal route

* use the new style for the start page

* warn when referencing a non-existent route

* add the repeater!

this uses `enterContext` and `exitContext` to iterate over the existing
data and render the contents statically multiple times on the backend.
the frontend will then manage the buttons and re-ordering, in such a way
that the data will be posted properly to the server.

* add a sample route 'addresses' that uses the repeater

* locale changes

* add a spec for the context helpers

* properly implement `isFirstError(relname)`

* properly use the errorState session key

* expand the contextHelpers spec

* clear the error state on any successful form submission

* behave better when req.body has falsey values

* comment

* rm a test for an interface that doesn't exist anymore

* rm a console.log

* properly propagate the errors from the session

* remove validateRouteData

we will want something similar to this, but the current implementation
is not workable.

* rm data.helpers.spec.js

* add in `mockSession`, a new way to mock the session

* protect against cases in tests where there is no session

* use console.warn for warnings

* rely on res.locals.firstError in `isFirstError`

* lint

* add a test for route.render()

* add a test for pad

* rm useless conditional

* use default arguments for pad

* fix the error list links

* better autofocus behaviour

* add unique identifiers and styling to repeaters

* allow pressing spacebar to activate the repeat and remove-repeat links

* standardize on `client.js` for client javascript

and have it automatically loaded if the file exists

* update repeater.js to es6 style

* delete an overly-brittle test

* better keyboard navigation behaviour

* move polyfill scripts to assets/js/

* move the only remaining nunjucks script to googleAnalytics.njk

* add assets/js/ to the default module path for client js

* add translations for the addresses page and fix confirmation

* remove the push hooks - we will rely on CI

* disable import-checking in eslint

bad imports should be caught immediately in testing anyways

* use modern variable scoping in repeater.js

* use template strings in repeater.js

* rename a confusingly-named helper function

* properly set up the webpack import resolver for eslint

* suppress the lgtm bot warning for document.write

* fix and simplify the logic of details-polyfill-detect

* use <button>s for the add / remove buttons